Remove redundant file
